Github Action to build on Windows, Ubuntu, and macOS
Builds the main library and PowerShell modules in separate workflows,
and uploads artifacts from both.

Library and Example are built and run on Windows, Ubuntu, and macOS.

PowerShell module is tested by importing the module on Windows
PowerShell and PowerShell Core on Windows, Ubuntu, and macOS.

Artifacts for both are built on Windows, but this isn't a requirement-
and OS would function identically.